+++ react/lib/module-declaration.ts
@@ -1,12 +1,17 @@
-import type { JBModalWebComponent } from 'jb-modal';
+import type { JBModalWebComponent } from "jb-modal";
 
 declare module "react" {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace JSX {
     interface IntrinsicElements {
-      'jb-modal': JBModalType;
+      "jb-modal": JBModalType;
     }
     interface JBModalType extends React.DetailedHTMLProps<React.HTMLAttributes<JBModalWebComponent>, JBModalWebComponent> {
+      label?: string;
+      description?: string;
+      autoCloseOnBackgroundClick?: boolean;
+      autoCloseOnEscape?: boolean;
+      "is-open"?: boolean | "true" | "false";
     }
   }
 }