+++ web-component/lib/render.ts
@@ -3,7 +3,7 @@ export function renderHTML(): string {
   <div class="jb-modal-web-component" part="component-wrapper">
     <div class="modal-background" part="background" aria-hidden="true"></div>
     <div class="modal-content-wrapper">
-        <div class="modal-content" part="content-box">
+        <div class="modal-content" part="content-box" tabindex="-1">
             <slot name="header"></slot>
             <slot name="content"><slot></slot></slot>
             <slot name="footer"></slot>